Fix warnings and non-standard code

+++ include/EverydayTools/Algorithm/Sort/InsertionSort.h
@@ -5,8 +5,8 @@ namespace edt
     template<typename T>
     void InsertionSort(T* arr, size_t size) {
         if (size > 1) {
-            for (int i = 1; i < size; ++i) {
-                for (int j = i; j > 0; --j) {
+            for (size_t i = 1; i < size; ++i) {
+                for (size_t j = i; j > 0; --j) {
                     T& a = arr[j];
                     T& b = arr[j - 1];
                     if (a < b) {
